requirements.txt
django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
pillow==11.1.0
python-decouple==3.8
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
django-allauth==65.3.1
django-rest-auth==0.9.5
django-filter==24.3
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
